move pca functions to pc.py
