+++ suite2p/registration/__init__.py
@@ -1,3 +1,3 @@
-from .metrics import get_pc_metrics
+from .pc import get_pc_metrics
 from .main import register_binary
 from .zalign import compute_zpos
